+++ NEWS.md
@@ -1,5 +1,7 @@
 # paletteer (development version)
 
+* added `scale_color_paletteer_binned()` to support binned scales.
+
 # paletteer 1.0.0
 
 * All front facing functions now take a single `palette` argument of the format `package::palette` instead of the  previous two arguments package and palette.